pyproject.toml
[tool.poetry]
name = "umbrella-policy-python-sdk"
version = "0.1.0"
description = "A simple API wrapper for the Cisco Umbrella Policies v2 API"
authors = ["Chris Doucette <chrisdoucette15@gmail.com>"]
readme = "README.md"
packages = [{include = "umbrella_policy_python_sdk"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
requests-oauthlib = "^1.3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
